Update: Header, Navigation, Promo
Add Navigation component

Correction: Header(old) -> Promo(new) & Navbar(old) -> Header(new)
