+++ src/components/App/App.jsx
@@ -1,4 +1,3 @@
-import Navbar from "../Navbar/Navbar";
 import Header from "../Header/Header";
 import Main from "../Main/Main";
 import Footer from "../Footer/Footer";
@@ -8,10 +7,7 @@ const App = () => {
   return (
     <div className='App'>
       <Header />
-      <Navbar />
-
       <Main />
-
       <Footer />
     </div>
   )
