+++ src/components/App/app.css
@@ -51,7 +51,7 @@ li {
 /*}*/
 
 
-/*Navbar Menu Opening Animation:*/
+/*Navigation Menu Opening Animation:*/
 /*.scale-up-center {*/
 /*    -webkit-animation: scale-up-center 0.4s cubic-bezier(0.390, 0.575, 0.565, 1.000) both;*/
 /*    animation: scale-up-center 0.4s cubic-bezier(0.390, 0.575, 0.565, 1.000) both;*/
