+++ src/components/Header/Header.jsx
@@ -1,25 +1,32 @@
-import image from '../../assets/images/header/image.png'
+import React, { useState } from 'react'
+import logo from '../../assets/images/header/logo.svg'
+
+import {AiOutlineMenu, AiOutlineClose} from "react-icons/ai";
+import {GrClose} from 'react-icons/gr'
+import Navigation from "../Navigation/Navigation";
 import './header.css'
 
-const Text = 'Учебный проект студента факультета \n Веб-разработки.'
+
 const Header = () => {
   return (
     <header className='header'>
-      <div className="header__wrapper wrapper">
-        <div className="header__content content">
-          <h1 className="content_heading new-line">
-            {Text}
-          </h1>
-          <p>
-            Листайте ниже, чтобы узнать больше про этот проект и его создателя.
-          </p>
-          <button type='button'>
-            Узнать больше
-          </button>
+      <div className="header__wrapper">
+        <div className="header__logo">
+          <img src={logo} alt="Logo"/>
         </div>
-        <div className="header__image image">
-          <img src={image} alt="Planet Earth Image"/>
-        </div>
+        <Navigation />
+
+        {/*  Smaller Devices */}
+        {/*<div className="header__menu">*/}
+        {/*  {toggleMenu ?*/}
+        {/*    <AiOutlineClose size={22} color='#fff' onClick={() => setToggleMenu(false)} />*/}
+        {/*    :*/}
+        {/*    <AiOutlineMenu color='#fff' size={28} onClick={() => setToggleMenu(true)} />*/}
+        {/*  }*/}
+        {/*  {toggleMenu && (*/}
+        {/*    <Menu />*/}
+        {/*  )}*/}
+        {/*</div>*/}
       </div>
     </header>
   )
