--- src/components/NavTab/NavTab.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import './nav-tab.css'
-const NavTab = () => {
-  return (
-    <div>NavTab</div>
-  )
-}
-export default NavTab
