+++ src/components/Promo/Promo.jsx
@@ -1,7 +1,28 @@
 import './promo.css'
+import image from "../../assets/images/header/image.png";
+
+const Text = 'Учебный проект студента факультета \n Веб-разработки.'
+
 const Promo = () => {
   return (
-    <div>Promo</div>
+    <section className='promo'>
+      <div className="promo__wrapper wrapper">
+        <div className="promo__content content">
+          <h1 className="content_heading new-line">
+            {Text}
+          </h1>
+          <p>
+            Листайте ниже, чтобы узнать больше про этот проект и его создателя.
+          </p>
+          <button type='button'>
+            <a href="#about-project">Узнать больше</a>
+          </button>
+        </div>
+        <div className="promo__image image">
+          <img src={image} alt="Planet Earth Image"/>
+        </div>
+      </div>
+    </section>
   )
 }
 export default Promo
