+++ src/components/Promo/promo.css
@@ -1,0 +1,137 @@
+.promo {
+    position: relative;
+    width: 100%;
+    background: var(--color-bg  );
+    font-family: var(--font-family);
+}
+
+.promo__wrapper {
+    padding: 0 4.37rem;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+.promo__content h1 {
+    font-family: var(--font-family);
+    font-size: 50px;
+    font-weight: 400;
+    line-height: 58px;
+
+    margin: calc(4.69rem + 4.625rem) 0 2.25rem;
+    max-width: 640px;
+}
+
+.new-line {
+    white-space: pre-line;
+}
+
+.promo__content p {
+    opacity: .5;
+    font-family: var(--font-family);
+    font-size: 13px;
+    font-weight: 400;
+    line-height: normal;
+
+    max-width: 310px;
+}
+
+.promo__content button {
+    margin: 9.12rem 0 4.69rem;
+    padding: 0.62rem 1.25rem;
+    background: transparent;
+    outline: none;
+    border-radius: 0.1875rem;
+    border: 1px solid var(--color-text-light);
+    opacity: .6;
+
+    color: var(--color-text-light);
+    font-family: var(--font-family);
+    font-size: 12px;
+    font-weight: 500;
+    line-height: 16px;
+    cursor: pointer;
+    transition: var(--transition-on-hover);
+}
+
+.promo__content button:hover {
+    background: var(--color-green);
+    color: var(--color-text-dark);
+    outline: none;
+    border: 1px solid var(--color-green);
+    opacity: 1;
+}
+
+.promo__image {
+    /*margin: calc(4.69rem + 4.625rem) 0 9.06rem 11.88rem;*/
+    margin: calc(4.69rem + 4.625rem) 0 9.06rem;
+}
+
+@media screen and (max-width: 768px) {
+    .promo__wrapper {
+        padding: 0 2.75rem;
+        display: flex;
+        flex-direction: column-reverse;
+        justify-content: center;
+        align-items: center;
+    }
+
+    .new-line {
+        white-space: normal;
+    }
+
+    .promo__image {
+        margin: calc(4.625rem + 8.75rem) 0 0;
+    }
+
+    .promo__content {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        text-align: center;
+    }
+
+    .promo__content h1 {
+        font-size: 40px;
+        line-height: 50px;
+        margin: 3rem auto 1rem;
+    }
+
+    .promo__content p {
+        max-width: 260px;
+    }
+
+    .promo__content button {
+        margin: 5.75rem 0 4.75rem;
+    }
+
+}
+
+@media screen and (max-width: 320px) {
+    .promo__wrapper {
+        padding: 0 0.63rem;
+    }
+
+    .promo__content h1 {
+        font-size: 30px;
+        line-height: 38px;
+        letter-spacing: -1.2px;
+        margin: 2.5rem auto 1.5rem;
+        width: 100%;
+    }
+
+    .promo__content button {
+        margin: 4.8rem 0 1.87rem;
+    }
+
+    .promo__image {
+        margin: calc(4.625rem + 4.38rem) 0 0;
+
+    }
+
+    .promo__image img {
+        /*margin: 140px 0 0;*/
+        max-width: 210px;
+    }
+}
